Add files via upload
Problem statement: This dataset includes descriptions of hypothetical samples 
corresponding to 23 species of gilled mushrooms in the Agaricus and 
Lepiota Family Mushroom. The objective is to develop a machine 
learning algorithm that will determine if a certain mushroom is edible 
or poisonous based on its features such as cap shape, cap color, gill 
color, etc. using different classifiers